ajuste na etapa3 e proto função de sombreamento

+++ etapa3/Terreno.h
@@ -7,7 +7,7 @@ class Terreno {
 
     int n;
     double rugosidade; 
-    double **terreno; 
+    int **terreno; 
     int max = 20, min = 0; 
 
     int gerarNmr(int min, int max);
@@ -23,7 +23,7 @@ class Terreno {
     void geradorDeMapa(int tamTerreno, double r);
     int consultarColuna();
     int consultarLinha();
-    double consultarAltitude(int l, int c);
+    int consultarAltitude(int l, int c);
     bool salvarTerreno(const std::string &nomeArquivo);
     bool lerTerreno(const std::string &nomeArquivo); 
 };
